Rename trust region 'management' to 'sampling'
